[Metrics UI] Add context.reason and alertOnNoData to Inventory alerts (#70260)
