+++ x-pack/plugins/infra/server/lib/alerting/common/types.ts
@@ -1,0 +1,33 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License;
+ * you may not use this file except in compliance with the Elastic License.
+ */
+
+export enum Comparator {
+  GT = '>',
+  LT = '<',
+  GT_OR_EQ = '>=',
+  LT_OR_EQ = '<=',
+  BETWEEN = 'between',
+  OUTSIDE_RANGE = 'outside',
+}
+
+export enum Aggregators {
+  COUNT = 'count',
+  AVERAGE = 'avg',
+  SUM = 'sum',
+  MIN = 'min',
+  MAX = 'max',
+  RATE = 'rate',
+  CARDINALITY = 'cardinality',
+  P95 = 'p95',
+  P99 = 'p99',
+}
+
+export enum AlertStates {
+  OK,
+  ALERT,
+  NO_DATA,
+  ERROR,
+}
