+++ x-pack/plugins/infra/server/lib/alerting/inventory_metric_threshold/inventory_metric_threshold_executor.ts
@@ -5,12 +5,20 @@
  */
 import { first, get } from 'lodash';
 import { i18n } from '@kbn/i18n';
+import moment from 'moment';
+import { toMetricOpt } from '../../../../common/snapshot_metric_i18n';
 import { AlertStates, InventoryMetricConditions } from './types';
 import { AlertExecutorOptions } from '../../../../../alerts/server';
 import { InventoryItemType, SnapshotMetricType } from '../../../../common/inventory_models/types';
 import { InfraBackendLibs } from '../../infra_types';
 import { METRIC_FORMATTERS } from '../../../../common/formatters/snapshot_metric_formats';
 import { createFormatter } from '../../../../common/formatters';
+import {
+  buildErrorAlertReason,
+  buildFiredAlertReason,
+  buildNoDataAlertReason,
+  stateToAlertMessage,
+} from '../common/messages';
 import { evaluateCondition } from './evaluate_condition';
 
 interface InventoryMetricThresholdParams {
@@ -18,13 +26,20 @@ interface InventoryMetricThresholdParams {
   filterQuery: string | undefined;
   nodeType: InventoryItemType;
   sourceId?: string;
+  alertOnNoData?: boolean;
 }
 
 export const createInventoryMetricThresholdExecutor = (
   libs: InfraBackendLibs,
   alertId: string
 ) => async ({ services, params }: AlertExecutorOptions) => {
-  const { criteria, filterQuery, sourceId, nodeType } = params as InventoryMetricThresholdParams;
+  const {
+    criteria,
+    filterQuery,
+    sourceId,
+    nodeType,
+    alertOnNoData,
+  } = params as InventoryMetricThresholdParams;
 
   const source = await libs.sources.getSourceConfiguration(
     services.savedObjectsClient,
@@ -48,26 +63,56 @@ export const createInventoryMetricThresholdExecutor = (
     const isNoData = results.some((result) => result[item].isNoData);
     const isError = results.some((result) => result[item].isError);
 
-    if (shouldAlertFire) {
+    const nextState = isError
+      ? AlertStates.ERROR
+      : isNoData
+      ? AlertStates.NO_DATA
+      : shouldAlertFire
+      ? AlertStates.ALERT
+      : AlertStates.OK;
+
+    let reason;
+    if (nextState === AlertStates.ALERT) {
+      reason = results
+        .map((result) => {
+          if (!result[item]) return '';
+          const resultWithVerboseMetricName = {
+            ...result[item],
+            metric: toMetricOpt(result[item].metric)?.text || result[item].metric,
+          };
+          return buildFiredAlertReason(resultWithVerboseMetricName);
+        })
+        .join('\n');
+    }
+    if (alertOnNoData) {
+      if (nextState === AlertStates.NO_DATA) {
+        reason = results
+          .filter((result) => result[item].isNoData)
+          .map((result) => buildNoDataAlertReason(result[item]))
+          .join('\n');
+      } else if (nextState === AlertStates.ERROR) {
+        reason = results
+          .filter((result) => result[item].isError)
+          .map((result) => buildErrorAlertReason(result[item].metric))
+          .join('\n');
+      }
+    }
+    if (reason) {
       alertInstance.scheduleActions(FIRED_ACTIONS.id, {
         group: item,
-        item,
-        valueOf: mapToConditionsLookup(results, (result) =>
+        alertState: stateToAlertMessage[nextState],
+        reason,
+        timestamp: moment().toISOString(),
+        value: mapToConditionsLookup(results, (result) =>
           formatMetric(result[item].metric, result[item].currentValue)
         ),
-        thresholdOf: mapToConditionsLookup(criteria, (c) => c.threshold),
-        metricOf: mapToConditionsLookup(criteria, (c) => c.metric),
+        threshold: mapToConditionsLookup(criteria, (c) => c.threshold),
+        metric: mapToConditionsLookup(criteria, (c) => c.metric),
       });
     }
 
     alertInstance.replaceState({
-      alertState: isError
-        ? AlertStates.ERROR
-        : isNoData
-        ? AlertStates.NO_DATA
-        : shouldAlertFire
-        ? AlertStates.ALERT
-        : AlertStates.OK,
+      alertState: nextState,
     });
   }
 };
