+++ x-pack/plugins/infra/server/lib/alerting/inventory_metric_threshold/register_inventory_metric_threshold_alert_type.ts
@@ -35,6 +35,7 @@ export const registerMetricInventoryThresholdAlertType = (libs: InfraBackendLibs
           schema.string({ validate: validateIsStringElasticsearchJSONFilter })
         ),
         sourceId: schema.string(),
+        alertOnNoData: schema.maybe(schema.boolean()),
       },
       { unknowns: 'allow' }
     ),
