+++ x-pack/plugins/infra/server/lib/alerting/metric_threshold/types.ts
@@ -3,18 +3,11 @@
  * or more contributor license agreements. Licensed under the Elastic License;
  * you may not use this file except in compliance with the Elastic License.
  */
-
 import { Unit } from '@elastic/datemath';
-export const METRIC_THRESHOLD_ALERT_TYPE_ID = 'metrics.alert.threshold';
+import { Comparator, AlertStates } from '../common/types';
 
-export enum Comparator {
-  GT = '>',
-  LT = '<',
-  GT_OR_EQ = '>=',
-  LT_OR_EQ = '<=',
-  BETWEEN = 'between',
-  OUTSIDE_RANGE = 'outside',
-}
+export { Comparator, AlertStates };
+export const METRIC_THRESHOLD_ALERT_TYPE_ID = 'metrics.alert.threshold';
 
 export enum Aggregators {
   COUNT = 'count',
@@ -26,13 +19,6 @@ export enum Aggregators {
   CARDINALITY = 'cardinality',
   P95 = 'p95',
   P99 = 'p99',
-}
-
-export enum AlertStates {
-  OK,
-  ALERT,
-  NO_DATA,
-  ERROR,
 }
 
 interface BaseMetricExpressionParams {
